Fixed error with resetting
Error occurs when reset button is clicked after selecting a member location to load their webpage. It resulted in the map returning NULL values for all states. We added some delays to ensure this won't happen and made a static list of number of members by state to ensure it does not glitch when resetting.